requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
StrEnum==0.4.15
django-oauth-toolkit==2.3.0
django[argon2]==23.1.0
